Fix silent uint8 overflow

+++ src/scripts/measure/helpers.py
@@ -280,7 +280,7 @@ def format_img(
         img = np.roll(img, shift=int(hshift * ny / 100), axis=1)
 
     if brightness:
-        img = (img * brightness / 100).astype(np.uint8)
+        img = (img.astype(np.float32) * brightness / 100).astype(np.uint8)
 
     if img.ndim == 2:
         # pygame requires 3 channels
